"Almost Done With This Shit"
"The WaveGenarator functions normally as well as the wave spawning. Now
we need to get the Server-Client communication done and also the client
visualization of the game Mechanics. Additionally we need to implement
the loop in which the enemies get moved and the towers shoot and the
life of the Player gets checked."